+++ custom/meteor-config.js
@@ -36,7 +36,6 @@ app_settings = {
   // if you want to add more colour palettes, just add their
   // name to this array and charts will be classed using this name
   palettes: ['Primary', 'Alternate', 'Grayscale'],
-
   help: 'https://github.com/globeandmail/chart-tool/tree/master/README.md',
 
   chart: {