+++ meteor/client/templates/chart-login/chart-login.js
@@ -1,0 +1,5 @@
+// Simply 'inherites' helpers from AccountsTemplates
+Template.chartLogin.helpers(AccountsTemplates.atSigninLinkHelpers);
+
+// Simply 'inherites' events from AccountsTemplates
+Template.chartLogin.events(AccountsTemplates.atSigninLinkEvents);