+++ meteor/lib/chart-tool/scripts/settings.js
@@ -16,11 +16,11 @@ var xAxis = {"display":true,"scale":"time","ticks":"auto","orient":"bottom","for
 var barHeight = 25;
 var barLabelOffset = 6;
 var bands = {"padding":0.12,"offset":0.06,"outerPadding":0.06};
-var source = {"prefix":"CHART TOOL","suffix":" » SOURCE:"};
-var social = {"facebook":{"label":"Facebook","icon":"https://cdnjs.cloudflare.com/ajax/libs/foundicons/3.0.0/svgs/fi-social-facebook.svg","redirect":"","appID":""},"twitter":{"label":"Twitter","icon":"https://cdnjs.cloudflare.com/ajax/libs/foundicons/3.0.0/svgs/fi-social-twitter.svg","via":"","hashtag":""},"email":{"label":"Email","icon":"https://cdnjs.cloudflare.com/ajax/libs/foundicons/3.0.0/svgs/fi-mail.svg"},"sms":{"label":"SMS","icon":"https://cdnjs.cloudflare.com/ajax/libs/foundicons/3.0.0/svgs/fi-telephone.svg"}};
-var image = {"enable":false,"base_path":"","expiration":30000,"filename":"thumbnail","extension":"png","thumbnailWidth":460};
-var embedJS = "http://beta.images.theglobeandmail.com/static/templates/tools/chart-tool/1.2.0/chart-tool.min.js";
-var embedCSS = "http://beta.images.theglobeandmail.com/static/templates/tools/chart-tool/1.2.0/chart-tool.min.css";
+var source = {"prefix":"The Buffalo News","suffix":" » SOURCE:"};
+var social = {"facebook":{"label":"Facebook","icon":"https://cdnjs.cloudflare.com/ajax/libs/foundicons/3.0.0/svgs/fi-social-facebook.svg","redirect":"","appID":""},"twitter":{"label":"Twitter","icon":"https://cdnjs.cloudflare.com/ajax/libs/foundicons/3.0.0/svgs/fi-social-twitter.svg","via":"","hashtag":""},"email":{"label":"Email","icon":"https://cdnjs.cloudflare.com/ajax/libs/foundicons/3.0.0/svgs/fi-mail.svg"}};
+var image = {"enable":false,"base_path":"charts/thumbnails","expiration":30000,"filename":"thumbnail","extension":"png","thumbnailWidth":460};
+var embedJS = "http://projects.buffalonews.com/charts/1.2.3/chart-tool.min.js";
+var embedCSS = "http://projects.buffalonews.com/charts/1.2.3/chart-tool.min.css";
 var config = {
 	CUSTOM: CUSTOM,
 	prefix: prefix$1,
@@ -48,7 +48,7 @@ var config = {
 };
 
 var name = "chart-tool";
-var version = "1.2.2";
+var version = "1.2.3";
 var buildVer = "0";
 
 app_version = version;
@@ -86,7 +86,6 @@ app_settings = {
   // if you want to add more colour palettes, just add their
   // name to this array and charts will be classed using this name
   palettes: ['Primary', 'Alternate', 'Grayscale'],
-
   help: 'https://github.com/globeandmail/chart-tool/tree/master/README.md',
 
   chart: {