+++ src/js/charts/components/header.js
@@ -28,20 +28,18 @@ export default function header(container, obj) {
 
   let qualifier;
 
-  if (obj.options.type === 'bar') {
-    if (obj.qualifier !== '' || obj.editable) {
-      qualifier = headerGroup
-        .append('div')
-        .attrs({
-          'class': () => {
-            let str = `${obj.prefix}chart_qualifier ${obj.prefix}chart_qualifier-bar`;
-            if (obj.editable) { str += ' editable-chart_qualifier'; }
-            return str;
-          },
-          'contentEditable': () => { return obj.editable ? true : false; }
-        })
-        .text(obj.qualifier);
-    }
+  if (obj.qualifier !== '' || obj.editable) {
+    qualifier = headerGroup
+      .append('div')
+      .attrs({
+        'class': () => {
+          let str = `${obj.prefix}chart_qualifier ${obj.prefix}chart_qualifier-bar`;
+          if (obj.editable) { str += ' editable-chart_qualifier'; }
+          return str;
+        },
+        'contentEditable': () => { return obj.editable ? true : false; }
+      })
+      .text(obj.qualifier);
   }
 
   let legend;