+++ src/js/charts/components/social.js
@@ -20,10 +20,6 @@ export default function social(node, obj) {
           obj.social[prop].url = constructMailURL(obj);
           obj.social[prop].popup = false;
           break;
-        case 'SMS':
-          obj.social[prop].url = constructSMSURL(obj);
-          obj.social[prop].popup = false;
-          break;
         default:
           console.log('Incorrect social item definition.');
       }
@@ -130,7 +126,7 @@ function windowPopup(url, width, height) {
 function constructFacebookURL(obj){
   const base = 'https://www.facebook.com/dialog/share?',
     redirect = obj.social.facebook.redirect;
-  let url = `app_id=${obj.social.facebook.appID}&amp;display=popup&amp;title=${obj.heading}&amp;description=From%20article${document.title}&amp;href=${window.location.href}&amp;redirect_uri=${redirect}`;
+  let url = `display=popup&amp;title=${obj.heading}&amp;description=From%20article${document.title}&amp;href=${window.location.href}&amp;redirect_uri=${redirect}`;
   if (obj.image && obj.image.enable) { url += `&amp;picture=${getThumbnailPath(obj)}`; }
   return `${base}${url}`;
 }
@@ -141,18 +137,11 @@ function constructMailURL(obj){
   return `${base}subject=${obj.heading}&amp;body=${obj.heading}${thumbnail}%0Afrom article: ${document.title}%0A${window.location.href}`;
 }
 
-function constructSMSURL(obj){
-  const base = 'sms:';
-  let url = `&body=Check%20out%20this%20chart: ${obj.heading}`;
-  if (obj.image && obj.image.enable) {  url += `%20${getThumbnailPath(obj)}`; }
-  return `${base}${url}`;
-}
-
 function constructTwitterURL(obj){
   const base = 'https://twitter.com/intent/tweet?',
     hashtag = (obj.social.twitter.hashtag) ? `&amp;hashtags=${obj.social.twitter.hashtag}` : '',
     via = (obj.social.twitter.via) ? `&amp;via=${obj.social.twitter.via}` : '';
-  let url = `url=${window.location.href}${via}&amp;text=${encodeURI(obj.heading)}${hashtag}`;
+  let url = `url=${window.location.href}${via}&amp;text=${encodeURI(obj.heading).replace(/'/g, "%27")}${hashtag}`;
   if (obj.image && obj.image.enable) {  url += `%20${getThumbnailPath(obj)}`; }
   return `${base}${url}`;
 }